prediction_prix_energie/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
matplotlib==3.8.4
seaborn==0.13.2
arch==6.3.0
yfinance==0.2.38
ta==0.11.0
joblib==1.4.0
pytest==8.1.1
python-dotenv==1.0.1
refinitiv-data # Nécessite une installation spécifique et des identifiants LSEG

